Refactor add custom form to a component, use react window for tokenlist
