+++ components/brave_wallet_ui/components/desktop/popup-modals/edit-visible-assets-modal/index.tsx
@@ -1,9 +1,10 @@
 import * as React from 'react'
-import { BraveWallet, WalletState } from '../../../../constants/types'
 import {
-  PopupModal,
-  AssetWatchlistItem,
-  SelectNetworkDropdown
+  BraveWallet,
+  WalletState
+} from '../../../../constants/types'
+import {
+  PopupModal
 } from '../..'
 import { NavButton } from '../../../extension'
 import { SearchBar } from '../../../shared'
@@ -11,35 +12,21 @@ import { getLocale } from '../../../../../common/locale'
 
 // Styled Components
 import {
-  StyledWrapper,
-  WatchlistScrollContainer,
+  Divider,
   LoadIcon,
   LoadingWrapper,
-  NoAssetText,
-  NoAssetRow,
   NoAssetButton,
-  FormWrapper,
-  InputLabel,
-  Input,
-  Divider,
-  ButtonRow,
-  ErrorText,
-  TopRow,
-  DividerText,
-  SubDivider,
-  DividerRow,
-  AdvancedButton,
-  AdvancedIcon,
-  FormRow,
-  FormColumn
+  NoAssetRow,
+  NoAssetText,
+  StyledWrapper,
+  TopRow
 } from './style'
 
-// Utils
-import Amount from '../../../../utils/amount'
-
 // hooks
-import { useAssetManagement, useLib, useTokenInfo } from '../../../../common/hooks'
+import { useAssetManagement } from '../../../../common/hooks'
 import { useSelector } from 'react-redux'
+import { AddCustomTokenForm } from '../../../shared/add-custom-token-form/add-custom-token-form'
+import { VirtualizedVisibleAssetsList } from './virtualized-visible-assets-list'
 
 export interface Props {
   onClose: () => void
@@ -47,48 +34,24 @@ export interface Props {
 
 const EditVisibleAssetsModal = ({ onClose }: Props) => {
   // redux
-  const {
-    addUserAssetError,
-    userVisibleTokensInfo,
-    selectedNetwork,
-    networkList,
-    fullTokenList
-  } = useSelector(({ wallet }: { wallet: WalletState }) => wallet)
-
+  const userVisibleTokensInfo = useSelector(({ wallet }: { wallet: WalletState }) => wallet.userVisibleTokensInfo)
+  const fullTokenList = useSelector(({ wallet }: { wallet: WalletState }) => wallet.fullTokenList)
+  const selectedNetwork = useSelector(({ wallet }: { wallet: WalletState }) => wallet.selectedNetwork)
+  const networkList = useSelector(({ wallet }: { wallet: WalletState }) => wallet.networkList)
   // custom hooks
-  const { getBlockchainTokenInfo } = useLib()
   const {
-    onAddCustomAsset,
     onUpdateVisibleAssets
   } = useAssetManagement()
-  const {
-    onFindTokenInfoByContractAddress,
-    foundTokenInfoByContractAddress
-  } = useTokenInfo(getBlockchainTokenInfo, userVisibleTokensInfo, fullTokenList, selectedNetwork)
 
   // Token List States
+  const [searchValue, setSearchValue] = React.useState<string>('')
   const [filteredTokenList, setFilteredTokenList] = React.useState<BraveWallet.BlockchainToken[]>([])
   const [updatedTokensList, setUpdatedTokensList] = React.useState<BraveWallet.BlockchainToken[]>([])
-  const [tokenDisplayAmount, setTokenDisplayAmount] = React.useState<number>(40)
+  const [tokenContractAddress, setTokenContractAddress] = React.useState<string>('')
 
   // Modal UI States
   const [showAddCustomToken, setShowAddCustomToken] = React.useState<boolean>(false)
-  const [showAdvancedFields, setShowAdvancedFields] = React.useState<boolean>(false)
   const [isLoading, setIsLoading] = React.useState<boolean>(false)
-  const [hasError, setHasError] = React.useState<boolean>(false)
-  const [showTokenIDRequired, setShowTokenIDRequired] = React.useState<boolean>(false)
-  const [showNetworkDropDown, setShowNetworkDropDown] = React.useState<boolean>(false)
-
-  // Form States
-  const [searchValue, setSearchValue] = React.useState<string>('')
-  const [tokenName, setTokenName] = React.useState<string>('')
-  const [tokenID, setTokenID] = React.useState<string>('')
-  const [tokenSymbol, setTokenSymbol] = React.useState<string>('')
-  const [tokenContractAddress, setTokenContractAddress] = React.useState<string>('')
-  const [tokenDecimals, setTokenDecimals] = React.useState<string>('')
-  const [coingeckoID, setCoingeckoID] = React.useState<string>('')
-  const [iconURL, setIconURL] = React.useState<string>('')
-  const [customAssetsNetwork, setCustomAssetsNetwork] = React.useState<BraveWallet.NetworkInfo>()
 
   // If a user removes all of their assets from the userVisibleTokenInfo list,
   // there is a check in the async/lib.ts folder that will still return the networks
@@ -103,15 +66,6 @@ const EditVisibleAssetsModal = ({ onClose }: Props) => {
       !userVisibleTokensInfo[0].visible
   }, [userVisibleTokensInfo])
 
-  const handleScroll = (event: React.UIEvent<HTMLDivElement>) => {
-    const top = event.currentTarget.scrollTop
-    const offset = event.currentTarget.offsetHeight
-    const height = event.currentTarget.scrollHeight
-    if (offset + top + 400 > height) {
-      setTokenDisplayAmount(tokenDisplayAmount + 40)
-    }
-  }
-
   React.useEffect(() => {
     if (isUserVisibleTokensInfoEmpty) {
       return
@@ -119,82 +73,22 @@ const EditVisibleAssetsModal = ({ onClose }: Props) => {
     setUpdatedTokensList(userVisibleTokensInfo)
   }, [userVisibleTokensInfo])
 
-  // Handle Form Input Changes
-  const handleTokenNameChanged = (event: React.ChangeEvent<HTMLInputElement>) => {
-    if (hasError) {
-      setHasError(false)
+  const nativeAsset = React.useMemo(() => {
+    return {
+      contractAddress: '',
+      decimals: selectedNetwork.decimals,
+      isErc20: false,
+      isErc721: false,
+      logo: selectedNetwork.iconUrls[0] ?? '',
+      name: selectedNetwork.symbolName,
+      symbol: selectedNetwork.symbol,
+      visible: true,
+      tokenId: '',
+      coingeckoId: '',
+      chainId: selectedNetwork.chainId,
+      coin: selectedNetwork.coin
     }
-    setTokenName(event.target.value)
-  }
-
-  const handleTokenIDChanged = (event: React.ChangeEvent<HTMLInputElement>) => {
-    if (hasError || showTokenIDRequired) {
-      setHasError(false)
-      setShowTokenIDRequired(false)
-    }
-    setTokenDecimals('0')
-    setTokenID(event.target.value)
-  }
-
-  const handleTokenSymbolChanged = (event: React.ChangeEvent<HTMLInputElement>) => {
-    if (hasError) {
-      setHasError(false)
-    }
-    setTokenSymbol(event.target.value)
-  }
-
-  const handleTokenAddressChanged = (event: React.ChangeEvent<HTMLInputElement>) => {
-    if (hasError) {
-      setHasError(false)
-    }
-    setTokenContractAddress(event.target.value)
-  }
-
-  const handleTokenDecimalsChanged = (event: React.ChangeEvent<HTMLInputElement>) => {
-    if (hasError) {
-      setHasError(false)
-    }
-    setTokenDecimals(event.target.value)
-  }
-
-  const handleCoingeckoIDChanged = (event: React.ChangeEvent<HTMLInputElement>) => {
-    if (hasError) {
-      setHasError(false)
-    }
-    setCoingeckoID(event.target.value)
-  }
-
-  const handleIconURLChanged = (event: React.ChangeEvent<HTMLInputElement>) => {
-    if (hasError) {
-      setHasError(false)
-    }
-    setIconURL(event.target.value)
-  }
-
-  const resetInputFields = () => {
-    setTokenName('')
-    setTokenContractAddress('')
-    setTokenSymbol('')
-    setTokenDecimals('')
-    setTokenID('')
-    setCoingeckoID('')
-    setIconURL('')
-  }
-
-  const nativeAsset = {
-    contractAddress: '',
-    decimals: selectedNetwork.decimals,
-    isErc20: false,
-    isErc721: false,
-    logo: selectedNetwork.iconUrls[0] ?? '',
-    name: selectedNetwork.symbolName,
-    symbol: selectedNetwork.symbol,
-    visible: true,
-    tokenId: '',
-    coingeckoId: '',
-    chainId: selectedNetwork.chainId,
-    coin: selectedNetwork.coin
-  }
+  }, [selectedNetwork])
 
   const tokenList = React.useMemo(() => {
     const userVisibleContracts = isUserVisibleTokensInfoEmpty
@@ -205,33 +99,30 @@ const EditVisibleAssetsModal = ({ onClose }: Props) => {
       ? fullTokenList
       : [nativeAsset, ...fullTokenList]
 
-    const filteredTokenRegistry = fullAssetsListPlusNativeToken.filter((token) => !userVisibleContracts.includes(token.contractAddress.toLowerCase()))
+    const filteredTokenRegistry = fullAssetsListPlusNativeToken
+      .filter((token) => !userVisibleContracts.includes(token.contractAddress.toLowerCase()))
 
     return isUserVisibleTokensInfoEmpty
       ? filteredTokenRegistry
       : [...userVisibleTokensInfo, ...filteredTokenRegistry]
-  }, [fullTokenList, userVisibleTokensInfo])
+  }, [isUserVisibleTokensInfoEmpty, fullTokenList, userVisibleTokensInfo, nativeAsset])
 
   React.useEffect(() => {
     // Added this timeout to throttle setting the list
     // to allow the modal to appear instantly
-    setTimeout(function () {
+    const timeoutId = setTimeout(function () {
       setFilteredTokenList(tokenList)
+      setIsLoading(false)
     }, 500)
-    if (!addUserAssetError) {
-      setShowAddCustomToken(false)
-      resetInputFields()
-    }
-    setHasError(addUserAssetError)
-    setIsLoading(false)
-  }, [tokenList, addUserAssetError])
+
+    return () => clearTimeout(timeoutId)
+  }, [tokenList])
 
   const filterWatchlist = React.useCallback((event: React.ChangeEvent<HTMLInputElement>) => {
     const search = event.target.value
     if (search === '') {
       setTimeout(function () {
         setFilteredTokenList(tokenList)
-        setTokenDisplayAmount(40)
       }, 100)
     } else {
       const filteredList = tokenList.filter((item) => {
@@ -248,63 +139,25 @@ const EditVisibleAssetsModal = ({ onClose }: Props) => {
     setSearchValue(search)
   }, [tokenList])
 
-  const onClickAddCustomToken = React.useCallback(() => {
-    if (!customAssetsNetwork) return
-
-    if (foundTokenInfoByContractAddress) {
-      if (foundTokenInfoByContractAddress.isErc721) {
-        let token = foundTokenInfoByContractAddress
-        token.tokenId = tokenID ? new Amount(tokenID).toHex() : ''
-        token.logo = token.logo ? token.logo : iconURL
-        token.chainId = customAssetsNetwork.chainId
-        setIsLoading(true)
-        onAddCustomAsset(token)
-        return
-      }
-      let foundToken = foundTokenInfoByContractAddress
-      foundToken.coingeckoId = coingeckoID !== '' ? coingeckoID : foundTokenInfoByContractAddress.coingeckoId
-      foundToken.logo = foundToken.logo ? foundToken.logo : iconURL
-      foundToken.chainId = customAssetsNetwork.chainId
-      onAddCustomAsset(foundToken)
-    } else {
-      const newToken: BraveWallet.BlockchainToken = {
-        contractAddress: tokenContractAddress,
-        decimals: Number(tokenDecimals),
-        isErc20: customAssetsNetwork.coin !== BraveWallet.CoinType.SOL && !tokenID,
-        isErc721: customAssetsNetwork.coin !== BraveWallet.CoinType.SOL && !!tokenID,
-        name: tokenName,
-        symbol: tokenSymbol,
-        tokenId: tokenID ? new Amount(tokenID).toHex() : '',
-        logo: iconURL,
-        visible: true,
-        coingeckoId: coingeckoID,
-        chainId: customAssetsNetwork.chainId,
-        coin: customAssetsNetwork.coin
-      }
-      onAddCustomAsset(newToken)
-    }
-    setIsLoading(true)
-  }, [tokenContractAddress, foundTokenInfoByContractAddress, customAssetsNetwork, iconURL, tokenDecimals, tokenName, tokenSymbol, tokenID, coingeckoID])
-
-  const findUpdatedTokenInfo = (token: BraveWallet.BlockchainToken) => {
+  const findUpdatedTokenInfo = React.useCallback((token: BraveWallet.BlockchainToken) => {
     return updatedTokensList.find((t) =>
       t.symbol.toLowerCase() === token.symbol.toLowerCase() &&
       t.contractAddress.toLowerCase() === token.contractAddress.toLowerCase() &&
       t.chainId === token.chainId)
-  }
+  }, [updatedTokensList])
 
-  const isUserToken = (token: BraveWallet.BlockchainToken) => {
+  const isUserToken = React.useCallback((token: BraveWallet.BlockchainToken) => {
     return updatedTokensList.some(t =>
     (
       t.contractAddress.toLowerCase() === token.contractAddress.toLowerCase() &&
       t.chainId === token.chainId &&
       t.symbol.toLowerCase() === token.symbol.toLowerCase())
     )
-  }
+  }, [updatedTokensList])
 
-  const isAssetSelected = (token: BraveWallet.BlockchainToken): boolean => {
+  const isAssetSelected = React.useCallback((token: BraveWallet.BlockchainToken): boolean => {
     return (isUserToken(token) && findUpdatedTokenInfo(token)?.visible) ?? false
-  }
+  }, [isUserToken, findUpdatedTokenInfo])
 
   const isCustomToken = React.useCallback((token: BraveWallet.BlockchainToken): boolean => {
     if (token.contractAddress === '') {
@@ -315,16 +168,14 @@ const EditVisibleAssetsModal = ({ onClose }: Props) => {
       .some(each => each.contractAddress.toLowerCase() === token.contractAddress.toLowerCase())
   }, [fullTokenList])
 
-  const addOrRemoveTokenFromList = (selected: boolean, token: BraveWallet.BlockchainToken) => {
+  const addOrRemoveTokenFromList = React.useCallback((selected: boolean, token: BraveWallet.BlockchainToken) => {
     if (selected) {
-      const addededToList = [...updatedTokensList, token]
-      return addededToList
+      return [...updatedTokensList, token]
     }
-    const removedFromList = updatedTokensList.filter((t) => t !== token)
-    return removedFromList
-  }
+    return updatedTokensList.filter((t) => t !== token)
+  }, [updatedTokensList])
 
-  const onCheckWatchlistItem = (key: string, selected: boolean, token: BraveWallet.BlockchainToken, isCustom: boolean) => {
+  const onCheckWatchlistItem = React.useCallback((key: string, selected: boolean, token: BraveWallet.BlockchainToken, isCustom: boolean) => {
     if (isUserToken(token)) {
       if (isCustom || token.contractAddress === '') {
         const updatedToken = selected ? { ...token, visible: true } : { ...token, visible: false }
@@ -336,32 +187,20 @@ const EditVisibleAssetsModal = ({ onClose }: Props) => {
         newList.splice(tokenIndex, 1, updatedToken)
         setUpdatedTokensList(newList)
       } else {
+        if (token.isErc721) setTokenContractAddress(token.contractAddress)
         setUpdatedTokensList(addOrRemoveTokenFromList(selected, token))
       }
     } else {
-      if (token.isErc721) {
-        setTokenContractAddress(token.contractAddress)
-        setShowTokenIDRequired(true)
-        setShowAddCustomToken(true)
-        return
-      }
       setUpdatedTokensList(addOrRemoveTokenFromList(selected, token))
     }
-  }
+  }, [isUserToken, updatedTokensList, addOrRemoveTokenFromList])
 
-  const toggleShowAddCustomToken = () => {
-    setShowAddCustomToken(!showAddCustomToken)
-  }
+  const toggleShowAddCustomToken = () => setShowAddCustomToken(prev => !prev)
 
-  const onClickSuggestAdd = () => {
+  const onClickSuggestAdd = React.useCallback(() => {
     setTokenContractAddress(searchValue)
     toggleShowAddCustomToken()
-  }
-
-  const onClickCancel = () => {
-    resetInputFields()
-    toggleShowAddCustomToken()
-  }
+  }, [searchValue, toggleShowAddCustomToken])
 
   const onRemoveAsset = React.useCallback((token: BraveWallet.BlockchainToken) => {
     const filterFn = (t: BraveWallet.BlockchainToken) => !(t.contractAddress.toLowerCase() === token.contractAddress.toLowerCase() && t.tokenId === token.tokenId)
@@ -371,70 +210,10 @@ const EditVisibleAssetsModal = ({ onClose }: Props) => {
     setFilteredTokenList(newFilteredTokenList)
   }, [updatedTokensList, filteredTokenList])
 
-  const isDecimalDisabled = React.useMemo((): boolean => {
-    return foundTokenInfoByContractAddress?.isErc721 ?? tokenID !== ''
-  }, [foundTokenInfoByContractAddress, tokenID])
-
-  React.useEffect(() => {
-    if (tokenContractAddress === '') {
-      resetInputFields()
-      return
-    }
-    onFindTokenInfoByContractAddress(tokenContractAddress)
-    if (foundTokenInfoByContractAddress) {
-      setTokenName(foundTokenInfoByContractAddress.name)
-      setTokenSymbol(foundTokenInfoByContractAddress.symbol)
-      setTokenDecimals(foundTokenInfoByContractAddress.decimals.toString())
-    }
-    if (foundTokenInfoByContractAddress?.isErc721) {
-      if (tokenID === '') {
-        setShowTokenIDRequired(true)
-        setShowAdvancedFields(true)
-      }
-    } else {
-      setShowTokenIDRequired(false)
-    }
-  }, [foundTokenInfoByContractAddress, tokenID, tokenContractAddress])
-
-  const buttonDisabled = React.useMemo((): boolean => {
-    return tokenName === '' ||
-      tokenSymbol === '' ||
-      (tokenDecimals === '0' && tokenID === '') ||
-      tokenDecimals === '' ||
-      tokenContractAddress === '' ||
-      !customAssetsNetwork ||
-      (customAssetsNetwork?.coin !== BraveWallet.CoinType.SOL &&
-        !tokenContractAddress.toLowerCase().startsWith('0x'))
-  }, [tokenName, tokenSymbol, tokenDecimals, tokenID, tokenContractAddress, customAssetsNetwork])
-
-  const onCloseModal = () => {
-    resetInputFields()
-    onClose()
-  }
-
-  const onClickDone = () => {
+  const onClickDone = React.useCallback(() => {
     onUpdateVisibleAssets(updatedTokensList)
-    onCloseModal()
-  }
-
-  const onToggleShowAdvancedFields = () => {
-    setShowAdvancedFields(!showAdvancedFields)
-  }
-
-  const onShowNetworkDropDown = () => {
-    setShowNetworkDropDown(true)
-  }
-
-  const onHideNetworkDropDown = React.useCallback(() => {
-    if (showNetworkDropDown) {
-      setShowNetworkDropDown(false)
-    }
-  }, [showNetworkDropDown])
-
-  const onSelectCustomNetwork = (network: BraveWallet.NetworkInfo) => {
-    setCustomAssetsNetwork(network)
-    onHideNetworkDropDown()
-  }
+    onClose()
+  }, [updatedTokensList, onUpdateVisibleAssets, onClose])
 
   return (
     <PopupModal
@@ -442,117 +221,25 @@ const EditVisibleAssetsModal = ({ onClose }: Props) => {
         ? getLocale('braveWalletWatchlistAddCustomAsset')
         : getLocale('braveWalletAccountsEditVisibleAssets')
       }
-      onClose={onCloseModal}
+      onClose={onClose}
     >
       {showAddCustomToken &&
         <Divider />
       }
-      <StyledWrapper onClick={onHideNetworkDropDown}>
+
+      <StyledWrapper>
         {(filteredTokenList.length === 0 && searchValue === '') || isLoading ? (
           <LoadingWrapper>
             <LoadIcon />
           </LoadingWrapper>
         ) : (
           <>
-            {showAddCustomToken ? (
-              <FormWrapper>
-                <InputLabel>{getLocale('braveWalletSelectNetwork')}</InputLabel>
-                <SelectNetworkDropdown
-                  selectedNetwork={customAssetsNetwork}
-                  onClick={onShowNetworkDropDown}
-                  showNetworkDropDown={showNetworkDropDown}
-                  onSelectCustomNetwork={onSelectCustomNetwork}
-                />
-                <FormRow>
-                  <FormColumn>
-                    <InputLabel>{getLocale('braveWalletWatchListTokenName')}</InputLabel>
-                    <Input
-                      value={tokenName}
-                      onChange={handleTokenNameChanged}
-                    />
-                  </FormColumn>
-                  <FormColumn>
-                    <InputLabel>{getLocale('braveWalletWatchListTokenAddress')}</InputLabel>
-                    <Input
-                      value={tokenContractAddress}
-                      onChange={handleTokenAddressChanged}
-                    />
-                  </FormColumn>
-                </FormRow>
-                <FormRow>
-                  <FormColumn>
-                    <InputLabel>{getLocale('braveWalletWatchListTokenSymbol')}</InputLabel>
-                    <Input
-                      value={tokenSymbol}
-                      onChange={handleTokenSymbolChanged}
-                    />
-                  </FormColumn>
-                  <FormColumn>
-                    <InputLabel>{getLocale('braveWalletWatchListTokenDecimals')}</InputLabel>
-                    <Input
-                      value={tokenDecimals}
-                      onChange={handleTokenDecimalsChanged}
-                      disabled={isDecimalDisabled}
-                      type='number'
-                    />
-                  </FormColumn>
-                </FormRow>
-                <DividerRow>
-                  <AdvancedButton onClick={onToggleShowAdvancedFields}>
-                    <DividerText>{getLocale('braveWalletWatchListAdvanced')}</DividerText>
-                  </AdvancedButton>
-                  <AdvancedButton onClick={onToggleShowAdvancedFields}>
-                    <AdvancedIcon rotated={showAdvancedFields} />
-                  </AdvancedButton>
-                </DividerRow>
-                <SubDivider />
-                {showAdvancedFields &&
-                  <>
-                    <InputLabel>{getLocale('braveWalletIconURL')}</InputLabel>
-                    <Input
-                      value={iconURL}
-                      onChange={handleIconURLChanged}
-                    />
-                    <InputLabel>{getLocale('braveWalletWatchListCoingeckoId')}</InputLabel>
-                    <Input
-                      value={coingeckoID}
-                      onChange={handleCoingeckoIDChanged}
-                    />
-                    {customAssetsNetwork?.coin !== BraveWallet.CoinType.SOL &&
-                      <>
-                        <InputLabel>{getLocale('braveWalletWatchListTokenId')}</InputLabel>
-                        <Input
-                          value={tokenID}
-                          onChange={handleTokenIDChanged}
-                          type='number'
-                          disabled={Number(tokenDecimals) > 0}
-                        />
-                      </>
-                    }
-                    {showTokenIDRequired &&
-                      <ErrorText>{getLocale('braveWalletWatchListTokenIdError')}</ErrorText>
-                    }
-                  </>
-                }
-                {hasError &&
-                  <ErrorText>{getLocale('braveWalletWatchListError')}</ErrorText>
-                }
-                <ButtonRow>
-                  <NavButton
-                    onSubmit={onClickCancel}
-                    text={getLocale('braveWalletButtonCancel')}
-                    buttonType='secondary'
-                  />
-                  <NavButton
-                    onSubmit={onClickAddCustomToken}
-                    text={getLocale('braveWalletWatchListAdd')}
-                    buttonType='primary'
-                    disabled={buttonDisabled}
-                  />
-                </ButtonRow>
-              </FormWrapper>
-            ) : (
-              <>
+            {showAddCustomToken
+              ? <AddCustomTokenForm
+                contractAddress={tokenContractAddress}
+                onHideForm={toggleShowAddCustomToken}
+              />
+              : <>
                 <SearchBar
                   value={searchValue}
                   placeholder={getLocale('braveWalletWatchListSearchPlaceholder')}
@@ -566,37 +253,31 @@ const EditVisibleAssetsModal = ({ onClose }: Props) => {
                     </NoAssetButton>
                   </TopRow>
                 }
-                <WatchlistScrollContainer onScroll={handleScroll}>
-                  <>
-                    {filteredTokenList.slice(0, tokenDisplayAmount).map((token) =>
-                      <AssetWatchlistItem
-                        key={`${token.contractAddress}-${token.symbol}-${token.chainId}-${token.tokenId}`}
-                        isCustom={isCustomToken(token)}
-                        token={token}
-                        networkList={networkList}
-                        onRemoveAsset={onRemoveAsset}
-                        isSelected={isAssetSelected(token)}
-                        onSelectAsset={onCheckWatchlistItem}
-                      />
-                    )}
-                    {filteredTokenList.length === 0 &&
-                      <NoAssetRow>
-                        {searchValue.toLowerCase().startsWith('0x') ? (
-                          <NoAssetButton onClick={onClickSuggestAdd}>{getLocale('braveWalletWatchListSuggestion').replace('$1', searchValue)}</NoAssetButton>
-                        ) : (
-                          <NoAssetText>{getLocale('braveWalletWatchListNoAsset')} {searchValue}</NoAssetText>
-                        )}
-                      </NoAssetRow>
-                    }
-                  </>
-                </WatchlistScrollContainer>
+                {filteredTokenList.length === 0
+                  ? <NoAssetRow>
+                      {searchValue.toLowerCase().startsWith('0x') ? (
+                        <NoAssetButton
+                          onClick={onClickSuggestAdd}>{getLocale('braveWalletWatchListSuggestion').replace('$1', searchValue)}</NoAssetButton>
+                      ) : (
+                        <NoAssetText>{getLocale('braveWalletWatchListNoAsset')} {searchValue}</NoAssetText>
+                      )}
+                  </NoAssetRow>
+                  : <VirtualizedVisibleAssetsList
+                    tokenList={filteredTokenList}
+                    networkList={networkList}
+                    isCustomToken={isCustomToken}
+                    onRemoveAsset={onRemoveAsset}
+                    isAssetSelected={isAssetSelected}
+                    onCheckWatchlistItem={onCheckWatchlistItem}
+                  />
+                }
                 <NavButton
                   onSubmit={onClickDone}
                   text={getLocale('braveWalletWatchListDoneButton')}
                   buttonType='primary'
                 />
               </>
-            )}
+            }
           </>
         )}
       </StyledWrapper>
