+++ components/brave_wallet_ui/components/desktop/popup-modals/edit-visible-assets-modal/style.ts
@@ -1,10 +1,13 @@
+// Copyright (c) 2022 The Brave Authors. All rights reserved.
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this file,
+// you can obtain one at http://mozilla.org/MPL/2.0/.
+
 import styled from 'styled-components'
-import { LoaderIcon, CaratStrongDownIcon } from 'brave-ui/components/icons'
+import { LoaderIcon } from 'brave-ui/components/icons'
 import { WalletButton } from '../../../shared/style'
 
-interface StyleProps {
-  rotated: boolean
-}
+export const tokenListHeight = 225
 
 export const StyledWrapper = styled.div`
   display: flex;
@@ -88,144 +91,10 @@ export const NoAssetButton = styled(WalletButton)`
   }
 `
 
-export const FormWrapper = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: flex-start;
-  width: 100%;
-  margin-top: 10px;
-`
-
-export const FormRow = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  width: 100%;
-`
-
-export const FormColumn = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  justify-content: center;
-  width: 49%;
-`
-
-export const InputLabel = styled.span`
-  font-family: Poppins;
-  font-size: 14px;
-  letter-spacing: 0.01em;
-  line-height: 20px;
-  color: ${(p) => p.theme.color.text02};
-  margin-bottom: 6px;
-`
-
-export const Input = styled.input`
-  outline: none;
-  width: 265px;
-  background-image: none;
-  background-color: ${(p) => p.theme.color.background02};
-  box-shadow: none;
-  border: ${(p) => `1px solid ${p.theme.color.interactive08}`};
-  border-radius: 4px;
-  font-family: Poppins;
-  font-style: normal;
-  font-size: 13px;
-  line-height: 20px;
-  letter-spacing: 0.01em;
-  padding: 10px;
-  margin-bottom: 8px;
-  color: ${(p) => p.theme.color.text01};
-  ::placeholder {
-    font-family: Poppins;
-    font-style: normal;
-    font-size: 12px;
-    letter-spacing: 0.01em;
-    color: ${(p) => p.theme.color.text03};
-    font-weight: normal;
-  }
-  :focus {
-    outline: none;
-  }
-  ::-webkit-inner-spin-button {
-    -webkit-appearance: none;
-    margin: 0;
-  }
-  ::-webkit-outer-spin-button {
-    -webkit-appearance: none;
-    margin: 0;
-  }
-`
-
 export const TopRow = styled.div`
   display: flex;
   flex-direction: row;
   align-items: flex-start;
   justify-content: flex-start;
   width: 100%;
-`
-
-export const ButtonRow = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  justify-content: flex-start;
-  width: 100%;
-  margin-top: 12px;
-`
-
-export const ErrorText = styled.span`
-  font-family: Poppins;
-  font-size: 12px;
-  line-height: 18px;
-  color: ${(p) => p.theme.color.errorText};
-  margin-bottom: 10px;
-`
-
-export const DividerRow = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-between;
-  width: 100%;
-  margin-top: 10px;
-`
-
-export const SubDivider = styled.div`
-  width: 100%;
-  height: 2px;
-  background-color: ${(p) => p.theme.color.divider01};
-  margin-bottom: 12px;
-`
-
-export const DividerText = styled.span`
-  font-family: Poppins;
-  font-size: 15px;
-  line-height: 20px;
-  letter-spacing: 0.04em;
-  font-weight: 600;
-  margin-bottom: 10px;
-  color: ${(p) => p.theme.color.text03};
-`
-
-export const AdvancedButton = styled(WalletButton)`
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-direction: row;
-  cursor: pointer;
-  outline: none;
-  background: none;
-  border: none;
-  padding: 0px;
-`
-
-export const AdvancedIcon = styled(CaratStrongDownIcon) <Partial<StyleProps>>`
-  width: 18px;
-  height: 18px;
-  color: ${(p) => p.theme.color.interactive07};
-  transform: ${(p) => p.rotated ? 'rotate(180deg)' : 'rotate(0deg)'};
-  margin-right: 10px;
 `
